+++ index.html
@@ -231,7 +231,7 @@
   </template>
 
   <!-- Первый скрипт -->
-  <script src="../js/main.js"></script>
+  <script src="../js/main.js" type="module"></script>
 
 </body>
 </html>
